pyproject.toml
[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Nif-Naf <Nif-Naf94@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
fastapi = "^0.109.2"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
python-dateutil = "^2.8.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
pydantic = {extras = ["email"], version = "^2.6.4"}
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
pytest-dotenv = "^0.5.2"
pre-commit = "^3.3.3"
isort = "^5.12.0"
black = "^23.11.0"
flake8 = "^6.1.0"
flake8-annotations = "^3.0.1"
flake8-annotations-coverage = "^0.0.6"
flake8-bandit = "^4.1.1"
flake8-broken-line = "^1.0.0"
flake8-bugbear = "^23.9.16"
flake8-builtins = "^2.2.0"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.14.0"
flake8-debugger = "^4.1.2"
flake8-eradicate = "^1.5.0"
flake8-executable = "^2.1.3"
flake8-fixme = "^1.1.1"
flake8-future-import = "^0.4.7"
flake8-pyi = "^23.11.0"
flake8-pytest = "^1.4"
flake8-pytest-style = "^1.7.2"
flake8-mutable = "^1.2.0"
flake8-string-format = "^0.3.0"
flake8-todo = "^0.7"
flake8-unused-arguments = "^0.0.13"
httpx = "^0.27.0"

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79

[tool.black]
line-length = 79

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q -p no:warnings"
asyncio_mode="auto"
log_cli = 1
log_cli_level = "INFO"
log_format = "%(levelname)s | Func: %(funcName)s | Line: %(lineno)d | %(message)s"
log_date_format = "%H:%M:%S"
testpaths = [
    "tests/unit",
    "tests/integration",
]
env_files = [
    ".test.env",
]
python_classes = [
    "Test*",
    "Testing*",
]
python_functions = [
    "test_*",
    "tests_*",
]
python_files = [
    "test_*",
    "tests_*",
    "testing_*"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
